Games/Firey Pong/pyproject.toml
[project]
name = "pong"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.6",
]
